Add caches to user show page. Download images for champions.
Added numerous caches to the user show page which are dependent on the most recent
game for a player. User show page now only makes 1 API call for most recent games, and if the first
game is the same as the cached version it skips all the subsequent calls that would normally
be made. Additionally changed the Presenter object to download champion images and save
them to the assets.